プレイヤーの状態を保持するPlayerStateクラスを追加 TaskStateをGameStateに移動
