--- Modules/TaskState.cs
@@ -1,31 +1,0 @@
-using BepInEx;
-using BepInEx.Configuration;
-using BepInEx.IL2CPP;
-using Hazel;
-using System;
-using HarmonyLib;
-using System.Collections.Generic;
-using System.IO;
-using UnityEngine;
-using UnhollowerBaseLib;
-using TownOfHost;
-namespace TownOfHost {
-    class TaskState {
-        public int AllTasksCount;
-        public int CompletedTasksCount;
-        public bool hasTasks;
-        public int RemainingTasksCount => AllTasksCount - CompletedTasksCount;
-        public bool doExpose => RemainingTasksCount <= main.SnitchExposeTaskLeft && hasTasks;
-        public bool isTaskFinished => RemainingTasksCount <= 0 && hasTasks;
-        public TaskState(int all, int completed) {
-            this.AllTasksCount = all;
-            this.CompletedTasksCount = completed;
-            this.hasTasks = true;
-        }
-        public TaskState() {
-            this.AllTasksCount = 0;
-            this.CompletedTasksCount = 0;
-            this.hasTasks = false;
-        }
-    }
-}
